Agents/requirements.txt
# A2A Protocol Dependencies
google-genai-a2a>=1.0.0  # Official Google A2A SDK
httpx>=0.24.0
uvicorn>=0.23.0
click>=8.1.0
pydantic>=2.0.0

# LangChain and LangGraph
langgraph>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-google-vertexai>=1.0.0

# Supabase
supabase>=2.0.0

# Other dependencies
python-dotenv>=1.0.0
requests>=2.31.0
